# swagger.yml
openapi: 3.0.0
info:
  title: "RP Flask REST API"
  description: "An API about graphs"
  version: "1.0.0"
servers:
  - url: "/api"
components:
...
  parameters:
    user_id:
      name: "user_id"
      description: "ID of the user"
      in: path
      required: True
      schema:
        type: "integer"
...
paths:
...
  /nodes/user/{user_id}:
    delete:
      tags:
        - "Nodes"
      operationId: "nodes.delete_all"
      summary: "delete all nodes and edges associated to a user"
      parameters:
        - $ref: "#/components/parameters/user_id"
      responses:
        "200":
          description: "Successfully deleted node"
        "404":
          description: "no nodes were found for the user"
...